// src/store/modules/actions-types.ts
export enum ActionsTypes {
  CACLULATE_BMR = "CACLULATE_BMR",
  CACLULATE_LBM = "CACLULATE_LBM",
  CACLULATE_PERFWEIGHT = "CACLULATE_PERFWEIGHT",
  CACLULATE_TDEE = "CACLULATE_TDEE",
  CACLULATE_BMI = "CACLULATE_BMI",
  CACLULATE_MACRO = "CACLULATE_MACRO",
  SET_NEWDATE = "CACLULATE_NEWDATE",
  SET_BASICINFO = "CACLULATE_BASICINFO",
  CACLULATE_GOAL = "CACLULATE_GOAL",
}